Move ansible-test units code into separate dir.
